validation completed of match
